requirements.txt
fastapi
uvicorn
python-dotenv
motor
bcrypt
PyJWT
pydantic
python-multipart